build_features now check for existence of the file

+++ benchtools/src/clustering.py
@@ -188,42 +188,48 @@ def build_features(path_data, nbatch, outname, path_label=None, outdir='../data'
         E and tau for the two principal jets, deltaR, mjj and number of 
         hadrons for a chuncksize of the events.
     """
-    # Creating the object to generate the dataframe
-    fb = generator(path_data,chunksize)
+    # Checking if the file hasn't been created before
+    filename=os.path.join(outdir,outname+'.csv').replace("\\","/")
+    if os.path.exists(filename):
+        print("A file with that name already exists")
+    # If not, continue with the creation of the file
+    else:    
+        # Creating the object to generate the dataframe
+        fb = generator(path_data,chunksize)
 
-    # Initiating the counter
-    batch_idx = 0
+        # Initiating the counter
+        batch_idx = 0
 
-    for batch in fb:
-        # A DataFrame to store the features 
-        df = pd.DataFrame(columns=['pT_j1', 'm_j1', 'eta_j1', 'phi_j1', 'E_j1', 'tau_21_j1', 'nhadrons_j1',
-                                'pT_j2', 'm_j2', 'eta_j2', 'phi_j2', 'E_j2', 'tau_21_j2', 'nhadrons_j2',
-                                'm_jj', 'deltaR_j12','n_hadrons', 'label'])
-        
-        # Checking the number of batch
-        if batch_idx < nbatch:
-            print('Part {}/{}'.format(batch_idx+1, nbatch))
-        elif batch_idx == nbatch:
-            print('Done')
-            break
+        for batch in fb:
+            # A DataFrame to store the features 
+            df = pd.DataFrame(columns=['pT_j1', 'm_j1', 'eta_j1', 'phi_j1', 'E_j1', 'tau_21_j1', 'nhadrons_j1',
+                                    'pT_j2', 'm_j2', 'eta_j2', 'phi_j2', 'E_j2', 'tau_21_j2', 'nhadrons_j2',
+                                    'm_jj', 'deltaR_j12','n_hadrons', 'label'])
             
-        data = batch
+            # Checking the number of batch
+            if batch_idx < nbatch:
+                print('Part {}/{}'.format(batch_idx+1, nbatch))
+            elif batch_idx == nbatch:
+                print('Done')
+                break
+                
+            data = batch
 
-        # Getting the label column if we are not using the RD DataFrame
-        if path_label != None:
-            label = ascii_column(path_label)
-            data = pd.concat([data, label.iloc[data.index]], axis=1)
+            # Getting the label column if we are not using the RD DataFrame
+            if path_label != None:
+                label = ascii_column(path_label)
+                data = pd.concat([data, label.iloc[data.index]], axis=1)
 
-        df = cluster_events(data)
-        
-        # Saving the DataFrame as csv for every batch  
-        savename = "".join((outname,'_{}')).format(batch_idx)
-        save_df(savename, outdir, df)
-        
-        batch_idx += 1 
-    # Merging in one file 
-    print('Merging files')
-    merge_files(outname, nbatch, outdir)
+            df = cluster_events(data)
+            
+            # Saving the DataFrame as csv for every batch  
+            savename = "".join((outname,'_{}')).format(batch_idx)
+            save_df(savename, outdir, df)
+            
+            batch_idx += 1 
+        # Merging in one file 
+        print('Merging files')
+        merge_files(outname, nbatch, outdir)
 
-    # Deleting the multiple files
-    delete_multifiles(outname, nbatch, outdir = '../data')
+        # Deleting the multiple files
+        delete_multifiles(outname, nbatch, outdir = '../data')
